+++ CHANGELOG.md
@@ -3,22 +3,29 @@
 
 ## [Unreleased]
 ### Added
+### Fixed
+### Changed
+### Deprecated
+### Removed
+### Security
+
+
+## 1.3.0
+
+### Added
 
 - Applicator class
 - Way to require explicit namespaces when applying permissions to datasets
 
-### Fixed
 ### Changed
 
 - PermissionSet.apply now uses the new Applicator class
 - Renamed core.list_namespace_handler to core.list_key_handler
 
-### Deprecated
 ### Removed
 
 - PermissionSet.handle_namespace (Replaced by Applicator class)
 
-### Security
 
 ## 1.2.0
 
@@ -26,6 +33,7 @@
 
 - namespace handlers
 - Support applying permissions to nested lists
+
 
 ## 1.1.0